Override breakpoints for iPhone 5
If it weren't for the 5's landscape being slightly smaller than the
default medium breakpoint, I wouldn't have wasted an hour figuring out
how to override those breakpoints.

So thank you, iPhone 5, for being so slightly off standard as to force
me to learn the order of sass includes to effectively override
foundation.